# PERF: speed up concat on Series by skipping unnecessary DataFrame creation