@qwhelan 2018-10-29
@jreback I was waiting for `asv` to finish before posting benchmarks, but I've updated the initial post with some basic benchmarks as that's taking longer than anticipated.