@qwhelan 2018-10-31
@TomAugspurger I implemented the `@classmethod` change for all but `_get_axis()` (which is used to get an axis instance, so shouldn't be a `@classmethod`).

There's decent existing test coverage but I added an explicit test of class vs instance invocations for all valid possible calls for good measure.