@TomAugspurger 2018-10-31
Thanks, LGTM, though you have a linting error: 
https://github.com/pandas-dev/pandas/pull/23404#issuecomment-433784500